requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.11.0
asyncpg==0.30.0
bcrypt==4.2.1
billiard==4.2.1
catboost==1.2.7
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
fonttools==4.56.0
graphviz==0.20.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.8
kombu==5.4.2
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
narwhals==1.28.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.1
pyTelegramBotAPI==4.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.18
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scipy==1.15.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
typer==0.14.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.2.3
uvicorn==0.32.1
vine==5.1.0
watchfiles==1.0.0
wcwidth==0.2.13
websockets==14.1
YooMoney==0.1.2
